requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
django-allauth
django-crispy-forms
idna
numpy
oauthlib
psycopg2-binary
pycparser
pyjwt[crypto]
python3-openid
requests
requests-oauthlib
sqlparse
tzdata
urllib3
